fields validation

+++ src/routes/tools.routes.ts
@@ -1,5 +1,6 @@
 import { Router } from 'express';
 import { getRepository } from 'typeorm';
+import * as Yup from 'yup';
 
 import Tool from '../models/Tool';
 import CreateToolService from '../services/CreateToolService';
@@ -27,6 +28,21 @@ toolsRouter.get('/', async (request, response) => {
 toolsRouter.post('/', async (request, response) => {
   const { title, link, description, tags } = request.body;
 
+  try {
+    const schema = Yup.object().shape({
+      title: Yup.string().required('Title is required'),
+      link: Yup.string().required('Link is required'),
+      description: Yup.string().required('Description is required'),
+      tags: Yup.array().required('Tags is required'),
+    });
+
+    await schema.validate(request.body, {
+      abortEarly: false,
+    });
+  } catch (err) {
+    return response.status(400).json(err.errors);
+  }
+
   const createTool = new CreateToolService();
 
   const tool = await createTool.execute({
@@ -43,6 +59,15 @@ toolsRouter.delete('/:id', async (request, response) => {
   const { id } = request.params;
 
   const toolsRepository = getRepository(Tool);
+
+  const repository = toolsRepository.findOne({
+    where: { id },
+  });
+
+  if (!repository) {
+    return response.status(400).json({ message: 'Repository does not exists' });
+  }
+
   toolsRepository.delete({ id });
 
   return response.status(204).send();
